+++ tera/src/tera.rs
@@ -474,7 +474,8 @@ impl Tera {
         self.register_filter("upper", crate::filters::upper);
         self.register_filter("lower", crate::filters::lower);
         self.register_filter("wordcount", crate::filters::wordcount);
-        self.register_filter("escape_html", crate::filters::escape);
+        self.register_filter("escape", crate::filters::escape);
+        self.register_filter("escape_html", crate::filters::escape_html);
         self.register_filter("escape_xml", crate::filters::escape_xml);
         self.register_filter("newlines_to_br", crate::filters::newlines_to_br);
         self.register_filter("pluralize", crate::filters::pluralize);